Create Classifying_Iris_Species
The code is about a hobby botanist that found the species of some iris flowers, she collected different measures in centimeters of each iris that are the length and width of the petals and sepals. She based in the data that an expert botanist identified as belonging to the species of setosa, versicolor and virginica. The code is going to identifie the measurements of the irises that had been discovered in order that we can predict the measurements of new species.
It is a classification problem because it wants to predict one of the species of the irises so the outputs will be the different species named as classes (we will have three classes). The desired output for an iris is the species of this flower. For a particular data point, the species it belongs to is called its label.